Changed Transfer to accept decimals for QuantityOrdered

Related work items: #31966

+++ IRMA/Projects/IrmaMobile/IrmaMobile/Domain/Models/TransferOrderItemModel.cs
@@ -7,7 +7,7 @@ namespace IrmaMobile.Domain.Models
 {
     public class TransferOrderItemModel
     {
-        public int QuantityOrdered { get; set; }
+        public decimal QuantityOrdered { get; set; }
         public int ItemKey { get; set; }
         public int QuantityUnit { get; set; }
         public decimal AdjustedCost { get; set; }
